backend/trading-journal/pyproject.toml
[project]
name = "trading-journal"
version = "0.1.0"
description = ""
authors = [
    {name = "Daeyan",email = "daeyansidi826@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
    "pytest (>=8.3.4,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "fastapi-cors (>=0.0.6,<0.0.7)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "psycopg[binary] (>=3.2.5,<4.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)"
]

[tool.poetry]
name = "trading-journal"
version = "0.1.0"
packages = [
    { include = "trading_journal", from = "src" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
